refactor room types
