+++ src/contexts/WargameProvider.tsx
@@ -3,7 +3,7 @@ import { WargameContext } from './WargameContext'
 import { XMPPService } from '../services/XMPPService'
 import { DataProvider } from 'react-admin'
 import { MockId } from '../types/wargame-d'
-import { roomTypeFactory } from '../roomTypes'
+import { roomTypeFactory } from '../services/roomTypes'
 
 interface WargameProviderProps {
   children: ReactNode
